requirements.txt
fastapi
uvicorn
websockets
google-genai
python-dotenv
google-auth
requests